+++ server/src/main/java/com/ticketer/utils/menu/dto/ComplexItem.java
@@ -4,11 +4,11 @@ import java.util.Map;
 
 public class ComplexItem {
     public String name;
-    public double basePrice;
+    public int basePrice;
     public boolean available;
     public Map<String, Side> sideOptions;
 
-    public ComplexItem(String name, double basePrice, boolean available, Map<String, Side> sideOptions) {
+    public ComplexItem(String name, int basePrice, boolean available, Map<String, Side> sideOptions) {
         this.name = name;
         this.basePrice = basePrice;
         this.available = available;
